refactor: Rename C++ header file extensions
Change C++ header file extensions to .hpp for simpler static analysis
tool setup.
