+++ msbuild/LibTromino.Test/LibTromino.Test.vcxproj
@@ -23,7 +23,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\test\libtromino\unittest_helper.h" />
+    <ClInclude Include="..\..\test\libtromino\unittest_helper.hpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
