+++ msbuild/TrominoPuzzle.Cli/TrominoPuzzle.Cli.vcxproj
@@ -112,12 +112,12 @@
     <ClCompile Include="..\..\src\cli\trmn_graph_windows.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\src\cli\cli_models.h" />
-    <ClInclude Include="..\..\src\cli\cli_options.h" />
-    <ClInclude Include="..\..\src\cli\init.h" />
-    <ClInclude Include="..\..\src\cli\trmn_graph.h" />
-    <ClInclude Include="..\..\src\cli\trmn_graph_vt.h" />
-    <ClInclude Include="..\..\src\cli\trmn_graph_windows.h" />
+    <ClInclude Include="..\..\src\cli\cli_models.hpp" />
+    <ClInclude Include="..\..\src\cli\cli_options.hpp" />
+    <ClInclude Include="..\..\src\cli\init.hpp" />
+    <ClInclude Include="..\..\src\cli\trmn_graph.hpp" />
+    <ClInclude Include="..\..\src\cli\trmn_graph_vt.hpp" />
+    <ClInclude Include="..\..\src\cli\trmn_graph_windows.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
