+++ src/Makefile.am
@@ -30,16 +30,16 @@ include_HEADERS += validation/tromino_validation.h
 lib_LTLIBRARIES += libtromino-validation.la
 
 tromino_SOURCES = \
-	cli/cli_models.h \
+	cli/cli_models.hpp \
 	cli/cli_options.cpp \
-	cli/cli_options.h \
+	cli/cli_options.hpp \
 	cli/init.cpp \
-	cli/init.h \
+	cli/init.hpp \
 	cli/main.cpp \
-	cli/params.h \
-	cli/trmn_graph.h \
+	cli/params.hpp \
 	cli/trmn_graph_vt.cpp \
-	cli/trmn_graph_vt.h
+	cli/trmn_graph_vt.hpp \
+	cli/trmn_graph.hpp
 tromino_CPPFLAGS = \
 	-I$(top_srcdir)/src/libtromino \
 	-I$(top_srcdir)/src/validation
@@ -49,15 +49,15 @@ tromino_LDADD = \
 bin_PROGRAMS += tromino
 
 libtromino_gfx2d_la_SOURCES = \
-	gfx2d/board.h \
-	gfx2d/step.h \
-	gfx2d/style.h \
+	gfx2d/board.hpp \
+	gfx2d/step.hpp \
+	gfx2d/style.hpp \
 	gfx2d/tromino_gfx2d.cpp \
-	gfx2d/tromino_gfx2d.h \
+	gfx2d/tromino_gfx2d.hpp \
 	gfx2d/view_model.cpp \
-	gfx2d/view_model.h \
+	gfx2d/view_model.hpp \
 	gfx2d/window.cpp \
-	gfx2d/window.h
+	gfx2d/window.hpp
 libtromino_gfx2d_la_CPPFLAGS = \
 	-I/usr/include/SDL2 \
 	-I$(top_srcdir)/src/libtromino
@@ -67,23 +67,23 @@ libtromino_gfx2d_la_LDFLAGS = \
 	-version-info $(LIBTROMINO_CURRENT):$(LIBTROMINO_REVISION):$(LIBTROMINO_AGE)
 libtromino_gfx2d_la_LIBADD = libtromino.la
 include_HEADERS += \
-	gfx2d/board.h \
-	gfx2d/step.h \
-	gfx2d/tromino_gfx2d.h \
-	gfx2d/view_model.h \
-	gfx2d/window.h
+	gfx2d/board.hpp \
+	gfx2d/step.hpp \
+	gfx2d/tromino_gfx2d.hpp \
+	gfx2d/view_model.hpp \
+	gfx2d/window.hpp
 lib_LTLIBRARIES += libtromino-gfx2d.la
 
 tromino2d_SOURCES = \
 	gui/callback.cpp \
-	gui/callback.h \
-	gui/cli_models.h \
+	gui/callback.hpp \
+	gui/cli_models.hpp \
 	gui/cli_options.cpp \
-	gui/cli_options.h \
+	gui/cli_options.hpp \
 	gui/init.cpp \
-	gui/init.h \
+	gui/init.hpp \
 	gui/main.cpp \
-	gui/params.h
+	gui/params.hpp
 tromino2d_CPPFLAGS = \
 	-I/usr/include/SDL2 \
 	-I$(top_srcdir)/src/libtromino \
