+++ src/cli/trmn_graph_windows.hpp
@@ -4,8 +4,8 @@
 
 // SPDX-License-Identifier: MIT
 
-#ifndef CLI_TRMN_GRAPH_WINDOWS_H
-#define CLI_TRMN_GRAPH_WINDOWS_H
+#ifndef CLI_TRMN_GRAPH_WINDOWS_HPP
+#define CLI_TRMN_GRAPH_WINDOWS_HPP
 
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
@@ -14,7 +14,7 @@
 
 #include "tromino.h"
 
-#include "trmn_graph.h"
+#include "trmn_graph.hpp"
 
 namespace tromino::cli::windows {
 
@@ -46,4 +46,4 @@ void use_wch(board_t& tromino_board, std::ostream& os) noexcept;
 
 } // namespace tromino::cli::windows
 
-#endif // CLI_TRMN_GRAPH_WINDOWS_H
+#endif // CLI_TRMN_GRAPH_WINDOWS_HPP
