+++ src/gfx2d/window.hpp
@@ -4,8 +4,8 @@
 
 // SPDX-License-Identifier: MIT
 
-#ifndef GFX2D_WINDOW_H
-#define GFX2D_WINDOW_H
+#ifndef GFX2D_WINDOW_HPP
+#define GFX2D_WINDOW_HPP
 
 #include <SDL2/SDL.h>
 
@@ -30,4 +30,4 @@ private:
 
 } // namespace tromino::gfx2d
 
-#endif // GFX2D_WINDOW_H
+#endif // GFX2D_WINDOW_HPP
