+++ src/gui/cli_options.cpp
@@ -4,9 +4,9 @@
 
 // SPDX-License-Identifier: MIT
 
-#include "cli_options.h"
+#include "cli_options.hpp"
 
-#include "params.h"
+#include "params.hpp"
 
 namespace tromino::tromino2d {
