+++ src/wasm/trmn/trmn.hpp
@@ -4,8 +4,8 @@
 
 // SPDX-License-Identifier: MIT
 
-#ifndef WASM_TRMN_TRMN_H
-#define WASM_TRMN_TRMN_H
+#ifndef WASM_TRMN_TRMN_HPP
+#define WASM_TRMN_TRMN_HPP
 
 #include <emscripten.h>
 
@@ -13,4 +13,4 @@ EMSCRIPTEN_KEEPALIVE extern "C" void playTromino(
     int const order, int const markX, int const markY,
     int const width) noexcept;
 
-#endif // WASM_TRMN_TRMN_H
+#endif // WASM_TRMN_TRMN_HPP
