+++ test/libtromino/unittest_helper.hpp
@@ -4,8 +4,8 @@
 
 // SPDX-License-Identifier: MIT
 
-#ifndef TEST_LIBTROMINO_UNITTEST_HELPER_H
-#define TEST_LIBTROMINO_UNITTEST_HELPER_H
+#ifndef TEST_LIBTROMINO_UNITTEST_HELPER_HPP
+#define TEST_LIBTROMINO_UNITTEST_HELPER_HPP
 
 #include <iostream>
 #include <vector>
@@ -54,4 +54,4 @@ void shim_add_tromino(
 
 void print_shim_step_vector(std::vector<ShimStep> const& steps) noexcept;
 
-#endif // TEST_LIBTROMINO_UNITTEST_HELPER_H
+#endif // TEST_LIBTROMINO_UNITTEST_HELPER_HPP
